+++ enrollment-server-onboarding-common/src/main/java/com/wultra/app/onboardingserver/common/database/entity/DocumentDataEntity.java
@@ -27,8 +27,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.proxy.HibernateProxy;
 
-import java.io.Serial;
-import java.io.Serializable;
 import java.util.Date;
 import java.util.Objects;
 
@@ -42,10 +40,7 @@ import java.util.Objects;
 @NoArgsConstructor
 @Entity
 @Table(name = "es_document_data")
-public class DocumentDataEntity implements Serializable {
-
-    @Serial
-    private static final long serialVersionUID = 7685715667785423079L;
+public class DocumentDataEntity {
 
     /***
      * Document data ID referenced as {@link DocumentVerificationEntity#uploadId}.