+++ docs/onboarding/Changelog.md
@@ -3,5 +3,6 @@
 
 ## 2.2.0
 
+- [#1425](https://github.com/wultra/enrollment-server/issues/1425) - Removed `Serializable` from JPA entities.
 - [#1592](https://github.com/wultra/enrollment-server/issues/1592) - Refactor state machine configuration, limit `EVENT_NEXT_STATE` and speed up processing.
 - [#1775](https://github.com/wultra/enrollment-server/issues/1775) - Migrated to Spring Boot 4 and Jackson 3.